src/main/java/ua/lviv/javaclub/githubcopilotdemo/user/UserService.java
package ua.lviv.javaclub.githubcopilotdemo.user;

import lombok.RequiredArgsConstructor;
import org.springframework.data.annotation.Transient;
import org.springframework.lang.NonNull;
import org.springframework.stereotype.Service;

import java.util.List;


@Service
@RequiredArgsConstructor
public class UserService {
    private final UserRepository userRepository;

    @Transient
    public User getUser(@NonNull final Long id) {
        return userRepository.findById(id)
                .orElseThrow(() -> new UserNotFoundException(id));
    }

    @Transient
    public List<User> getUsers() {
        return userRepository.findAllBy();
    }
}
